Bloglist 4.21

+++ controllers/blogs.js
@@ -42,10 +42,24 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const token = request.token
+    if (!token) {
+        return response.status(401).json({ error: 'token missing' })
+    }
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
     const blog = await Blog.findById(request.params.id)
+
     if (blog) {
-        await Blog.findByIdAndRemove(request.params.id)
-        response.status(204).end()
+        if (blog.user.toString() === decodedToken.id.toString()) {
+            await Blog.findByIdAndRemove(request.params.id)
+            response.status(204).end()
+        } else {
+            return response.status(401).json({ error: 'not user for blog id to delete' })
+        }
     } else {
         response.status(404).end()
     }
